Refactored file organization (components in folders); got correct eslint dependencies so it works in the project; changed prettier settings - this is making the diff look bigger than it actually is...
