+++ package.json
@@ -1,47 +1,51 @@
 {
-  "name": "brendan-crich-frontend-test",
-  "version": "0.0.0",
-  "main": "webpack.config.js",
-  "private": true,
-  "scripts": {
-    "start": "cross-env NODE_ENV=development webpack-dev-server --open",
-    "build": "cross-env NODE_ENV=production webpack"
-  },
-  "dependencies": {
-    "@material-ui/core": "^4.4.2",
-    "@material-ui/icons": "^4.4.1",
-    "classnames": "^2.2.6",
-    "eslint": "^6.4.0",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
-    "react-redux": "^7.1.1",
-    "react-router": "^5.0.1",
-    "react-router-dom": "^5.0.1",
-    "react-star-ratings": "^2.3.0",
-    "redux-saga": "^1.1.1",
-    "redux-thunk": "^2.3.0",
-    "truncate": "^2.1.0",
-    "whatwg-fetch": "^3.0.0"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.6.0",
-    "@babel/preset-env": "^7.6.0",
-    "@babel/preset-react": "^7.0.0",
-    "autoprefixer": "^9.6.1",
-    "babel-loader": "^8.0.6",
-    "clean-webpack-plugin": "^3.0.0",
-    "cross-env": "^6.0.0",
-    "css-loader": "^3.2.0",
-    "html-webpack-plugin": "^3.2.0",
-    "mini-css-extract-plugin": "^0.8.0",
-    "node-sass": "^4.12.0",
-    "postcss-loader": "^3.0.0",
-    "raw-loader": "^3.1.0",
-    "redux-devtools-extension": "^2.13.8",
-    "sass-loader": "^8.0.0",
-    "style-loader": "^1.0.0",
-    "webpack": "^4.40.2",
-    "webpack-cli": "^3.3.9",
-    "webpack-dev-server": "^3.8.1"
-  }
+    "name": "brendan-crich-frontend-test",
+    "version": "0.0.0",
+    "main": "webpack.config.js",
+    "private": true,
+    "scripts": {
+        "start": "cross-env NODE_ENV=development webpack-dev-server --open",
+        "build": "cross-env NODE_ENV=production webpack"
+    },
+    "dependencies": {
+        "@material-ui/core": "^4.4.2",
+        "@material-ui/icons": "^4.4.1",
+        "classnames": "^2.2.6",
+        "react": "^16.9.0",
+        "react-dom": "^16.9.0",
+        "react-redux": "^7.1.1",
+        "react-router": "^5.0.1",
+        "react-router-dom": "^5.0.1",
+        "react-star-ratings": "^2.3.0",
+        "redux-saga": "^1.1.1",
+        "redux-thunk": "^2.3.0",
+        "truncate": "^2.1.0",
+        "whatwg-fetch": "^3.0.0"
+    },
+    "devDependencies": {
+        "@babel/core": "^7.6.0",
+        "@babel/preset-env": "^7.6.0",
+        "@babel/preset-react": "^7.0.0",
+        "autoprefixer": "^9.6.1",
+        "babel-loader": "^8.0.6",
+        "clean-webpack-plugin": "^3.0.0",
+        "cross-env": "^6.0.0",
+        "css-loader": "^3.2.0",
+        "eslint": "^6.4.0",
+        "eslint-plugin-prettier": "^3.1.1",
+        "eslint-plugin-react": "^7.14.3",
+        "eslint-plugin-react-hooks": "^2.0.1",
+        "html-webpack-plugin": "^3.2.0",
+        "mini-css-extract-plugin": "^0.8.0",
+        "node-sass": "^4.12.0",
+        "postcss-loader": "^3.0.0",
+        "prettier": "^1.18.2",
+        "raw-loader": "^3.1.0",
+        "redux-devtools-extension": "^2.13.8",
+        "sass-loader": "^8.0.0",
+        "style-loader": "^1.0.0",
+        "webpack": "^4.40.2",
+        "webpack-cli": "^3.3.9",
+        "webpack-dev-server": "^3.8.1"
+    }
 }
