--- src/components/Filters.module.scss
@@ -1,40 +1,0 @@
-@import '../scss/variables';
-@import '../scss/extensions';
-:local .filters {
-    height: 80px;
-    position: relative;
-}
-:local .filtersRow {
-    position: absolute;
-    display: flex;
-    top: 50%;
-    transform: translateY(-50%);
-}
-:local .filterByLabel {
-    font-size: $font-size-label;
-    color: $color-grey;
-    height: 32px;
-    line-height: 1.8;
-    float: left;
-    margin-right: 30px;
-}
-:local .filterContainer {
-    margin-right: 30px;
-}
-:local .clearFiltersButton {
-    width: 151px;
-    height: 38px;
-    color: $color-primary;
-    font-size: $font-size-small;
-    border: 1px solid $color-primary;
-    border-radius: 1px;
-    cursor: pointer;
-    text-align: center;
-    line-height: 3.3;
-    float: right;
-    margin-top: 22px;
-
-    &:hover {
-        background-color: $color-light-grey;
-    }
-}
