+++ src/components/Header/Header.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import classNames from 'classnames';
+import { header, title } from './Header.module.scss';
+
+const Header = ({ children, headerTitle }) => {
+	return (
+		<div className={classNames(header, 'contentContainer')}>
+			<div className={title}>{headerTitle}</div>
+			{children}
+		</div>
+	);
+};
+// TODO proptypes
+export default Header;
