+++ src/components/Header/Header.module.scss
@@ -1,0 +1,25 @@
+@import '../../scss/variables';
+@import '../../scss/extensions';
+
+:local .header {
+	height: 194px;
+	margin-top: 36px;
+}
+:local .title {
+	font-family: $font-family-primary-light;
+	font-size: $font-size-large;
+	justify-content: center;
+	margin-bottom: 25px;
+}
+:local .description {
+	font-family: $font-family-primary-light;
+	color: $color-grey;
+	width: 650px;
+}
+:local .stickyTop {
+	position: fixed;
+	background: $color-white;
+	z-index: 10;
+	top: 0;
+	width: 100%;
+}
