+++ src/components/Loader/Loader.module.scss
@@ -1,0 +1,12 @@
+.loaderContainer {
+	position: relative;
+	background: rgba(255, 255, 255, 1);
+	width: 100%;
+	height: 48px;
+}
+.spinner {
+	position: absolute;
+	left: 50%;
+	top: 50%;
+	transform: translate(-50%, -50%);
+}
