--- src/components/Restaurant.module.scss
@@ -1,42 +1,0 @@
-@import '../scss/variables';
-
-:local .restaurantContainer {
-    width: 300px;
-    height: 430px;
-    margin: 0px 30px 80px 0px;
-    position: relative;
-}
-:local .restaurantImage {
-    width: 100%;
-    height: 228px;
-    background-color: $color-light-grey;
-    object-fit: cover;
-}
-:local .restaurantName {
-    color: $color-black;
-    margin-top: 20px;
-}
-:local .starRating {
-    position: relative;
-    margin-top: 10px;
-}
-:local .info {
-    margin-top: 15px;
-    font-size: $font-size-small;
-    color: $color-grey;
-}
-:local .learnMoreButton {
-    background: $color-primary;
-    a {
-        color: $color-white;
-    }
-    width: 100%;
-    height: 48px;
-    border-radius: 1px;
-    cursor: pointer;
-    font-size: $font-size-medium;
-    text-align: center;
-    line-height: 3.5;
-    position: absolute;
-    bottom: 0;
-}
