+++ src/components/Restaurant/Restaurant.module.scss
@@ -1,0 +1,42 @@
+@import '../../scss/variables';
+
+.restaurantContainer {
+	width: 300px;
+	height: 430px;
+	margin: 0px 30px 80px 0px;
+	position: relative;
+}
+.restaurantImage {
+	width: 100%;
+	height: 228px;
+	background-color: $color-light-grey;
+	object-fit: cover;
+}
+.restaurantName {
+	color: $color-black;
+	margin-top: 20px;
+}
+.starRating {
+	position: relative;
+	margin-top: 10px;
+}
+.info {
+	margin-top: 15px;
+	font-size: $font-size-small;
+	color: $color-grey;
+}
+.learnMoreButton {
+	background: $color-primary;
+	a {
+		color: $color-white;
+	}
+	width: 100%;
+	height: 48px;
+	border-radius: 1px;
+	cursor: pointer;
+	font-size: $font-size-medium;
+	text-align: center;
+	line-height: 3.5;
+	position: absolute;
+	bottom: 0;
+}
