--- src/components/RestaurantDetail.module.scss
@@ -1,35 +1,0 @@
-@import '../scss/variables';
-
-.ratingContainer {
-    position: relative;
-    transform: scale(1.5);
-    transform-origin: top left;
-}
-.locationDetails {
-    width: 100%;
-    height: 370px;
-    padding-top: 50px;
-}
-.priceText,
-.reviewCountText {
-    margin-top: 30px;
-    color: $color-grey;
-    font-family: $font-family-primary-light;
-}
-.reviewCountText {
-    margin-top: 60px;
-}
-.locationDetailsImageRow {
-    img {
-        width: 640px;
-        height: 228px;
-        object-fit: cover;
-    }
-}
-.addressText {
-    margin-top: 20px;
-    font-family: $font-family-primary-light;
-}
-.openNowContainer {
-    margin-top: 30px;
-}
