--- src/components/RestaurantsList.module.scss
@@ -1,24 +1,0 @@
-@import '../scss/variables';
-@import '../scss/extensions';
-
-:local .restaurantsList {
-    display: flex;
-    flex-flow: row wrap;
-}
-:local .title {
-    font-size: $font-size-title;
-    font-family: $font-family-primary-light;
-    margin-bottom: 40px;
-}
-:local .loadMoreButton {
-    width: 416px;
-    height: 48px;
-    color: $color-primary;
-    font-size: $font-size-medium;
-    border: 1px solid $color-primary;
-    border-radius: 1px;
-    cursor: pointer;
-    text-align: center;
-    line-height: 3.5;
-    margin: 0 auto;
-}
