+++ src/index.js
@@ -6,15 +6,15 @@ import './scss/base.scss';
 
 import { BrowserRouter, Route, Redirect, Switch } from 'react-router-dom';
 import store from './store';
-import App from './components/App.jsx';
-import Detail from './components/RestaurantDetail.jsx';
+import AllRestaurants from './components/AllRestaurants/AllRestaurants.jsx';
+import Detail from './components/RestaurantDetail/RestaurantDetail.jsx';
 
 render(
-    <Provider store={store}>
-        <BrowserRouter>
-            <Route exact path="/" component={App} />
-            <Route path="/detail/:id" component={Detail} />
-        </BrowserRouter>
-    </Provider>,
-    document.getElementById('root')
+	<Provider store={store}>
+		<BrowserRouter>
+			<Route exact path="/" component={AllRestaurants} />
+			<Route path="/detail/:id" component={Detail} />
+		</BrowserRouter>
+	</Provider>,
+	document.getElementById('root')
 );
